refactor list command into a new file
